Added kyc helper, removed membership helper
