+++ lib/kyc/buy_membership_helper.dart
@@ -1,0 +1,19 @@
+import 'package:commerciosdk/export.dart';
+import 'package:meta/meta.dart';
+
+/// Allows to easily create a BuyMembership and perform common related operations
+class BuyMembershipHelper {
+  /// Creates a BuyMembership from the given [wallet],
+  /// [membershipType] and [tsp] address.
+  static BuyMembership fromWallet({
+    @required Wallet wallet,
+    @required MembershipType membershipType,
+    @required String tsp,
+  }) {
+    return BuyMembership(
+      buyerDid: wallet.bech32Address,
+      membershipType: membershipType.value,
+      tsp: tsp,
+    );
+  }
+}
